Simplify logger creation
The previous logger creation was heavily dependent on the order modules
were imported.

The formatter also includes the pdb id so logging statements do not need
to explicitly include this, which makes logging less obnoxious.

This commit also helps separate logger configuration for each
application, which is work still to be done.